+++ src/components/add-review/add-review.jsx
@@ -1,105 +1,22 @@
-import React, {createRef} from "react";
+import React from "react";
 import PropTypes from "prop-types";
 import UserBlock from "../user-block/user-block.jsx";
-import {connect} from "react-redux";
-import {Operation as CommentOperation} from "../../reducer/comments/comments.js";
-import {CommentLength} from "../../utils/utils.js";
-import {AuthorizationStatus} from "../../reducer/user/user.js";
-import {getAuthStatus} from "../../reducer/user/selectors.js";
-import {Redirect} from "react-router-dom";
 
 class AddReview extends React.PureComponent {
   constructor(props) {
     super(props);
-
-    this.state = {
-      rating: null,
-      comment: ``,
-      isButtonDisabled: true,
-      readOnly: false,
-      errorText: ``,
-    };
-
-    this.buttonRef = createRef();
-
-    this.handleInputComment = this.handleInputComment.bind(this);
-    this.handleChangeRating = this.handleChangeRating.bind(this);
-    this.handleSendComment = this.handleSendComment.bind(this);
-  }
-
-  handleInputComment(event) {
-    this.setState({
-      comment: event.target.value,
-    });
-
-    this.checkDisableButton();
-  }
-
-  handleChangeRating(event) {
-    this.setState({
-      rating: +event.target.value,
-    });
-    this.checkDisableButton();
-  }
-
-  handleSendComment(event) {
-    event.preventDefault();
-
-    this.setState({
-      readOnly: true,
-      isButtonDisabled: true,
-      errorText: ``,
-    });
-
-    const onSuccess = () => {
-      this.setState({
-        readOnly: false,
-        isButtonDisabled: false,
-      });
-      this.props.history.push(`/dev-movie/${this.props.match.params.id}`);
-    };
-
-    const onError = (errorText) => {
-      this.setState({
-        readOnly: false,
-        isButtonDisabled: false,
-        errorText,
-      });
-    };
-
-    this.props.onSendComment({
-      movieId: this.props.match.params.id,
-      comment: this.state.comment,
-      rating: this.state.rating,
-    }, onSuccess, onError);
-  }
-
-  checkDisableButton() {
-    const commentLength = this.state.comment.length;
-    if (commentLength < CommentLength.MIN && commentLength >= CommentLength.MAX && !this.state.rating) {
-      this.setState({
-        isButtonDisabled: true,
-      });
-    } else {
-      this.setState({
-        isButtonDisabled: false,
-      });
-    }
   }
 
   renderErrorMessage() {
-    const {errorText} = this.state;
+    const {errorText} = this.props;
+
     return (
       <p className="add-review__text-error">{errorText}</p>
     );
   }
 
   render() {
-    const {statusAuth} = this.props;
-
-    if (statusAuth === AuthorizationStatus.NO_AUTH) {
-      return <Redirect to='/login'/>;
-    }
+    const {onInputComment, onChangeRating, onSendComment, isButtonDisabled, readOnly, errorText} = this.props;
 
     return (
       <section className="movie-card movie-card--full">
@@ -140,22 +57,22 @@ class AddReview extends React.PureComponent {
 
         <div className="add-review">
           <form
-            onSubmit={this.handleSendComment}
+            onSubmit={onSendComment}
             action="#"
             className="add-review__form"
           >
-            {this.state.errorText ? this.renderErrorMessage() : ``}
+            {errorText ? this.renderErrorMessage() : ``}
             <div className="rating">
               <div
                 className="rating__stars"
-                onChange={this.handleChangeRating}
+                onChange={onChangeRating}
               >
                 <input
                   className="rating__input"
                   id="star-1" type="radio"
                   name="rating"
                   value="1"
-                  disabled={this.state.readOnly}
+                  disabled={readOnly}
                 />
                 <label className="rating__label" htmlFor="star-1">Rating 1</label>
 
@@ -164,7 +81,7 @@ class AddReview extends React.PureComponent {
                   id="star-2" type="radio"
                   name="rating"
                   value="2"
-                  disabled={this.state.readOnly}
+                  disabled={readOnly}
                 />
                 <label className="rating__label" htmlFor="star-2">Rating 2</label>
 
@@ -173,7 +90,7 @@ class AddReview extends React.PureComponent {
                   id="star-3" type="radio"
                   name="rating"
                   value="3"
-                  disabled={this.state.readOnly}
+                  disabled={readOnly}
                 />
                 <label className="rating__label" htmlFor="star-3">Rating 3</label>
 
@@ -182,7 +99,7 @@ class AddReview extends React.PureComponent {
                   id="star-4" type="radio"
                   name="rating"
                   value="4"
-                  disabled={this.state.readOnly}
+                  disabled={readOnly}
                 />
                 <label className="rating__label" htmlFor="star-4">Rating 4</label>
 
@@ -191,7 +108,7 @@ class AddReview extends React.PureComponent {
                   id="star-5" type="radio"
                   name="rating"
                   value="5"
-                  disabled={this.state.readOnly}
+                  disabled={readOnly}
                 />
                 <label className="rating__label" htmlFor="star-5">Rating 5</label>
               </div>
@@ -199,18 +116,18 @@ class AddReview extends React.PureComponent {
 
             <div className="add-review__text">
               <textarea
-                onInput={this.handleInputComment}
+                onInput={onInputComment}
                 className="add-review__textarea"
                 name="review-text"
                 id="review-text"
                 placeholder="Review text"
-                readOnly={this.state.readOnly}
+                readOnly={readOnly}
               >
 
               </textarea>
               <div className="add-review__submit">
                 <button
-                  disabled={this.state.isButtonDisabled}
+                  disabled={isButtonDisabled}
                   className="add-review__btn"
                   type="submit"
                 >
@@ -227,29 +144,15 @@ class AddReview extends React.PureComponent {
 }
 
 AddReview.propTypes = {
-  match: PropTypes.shape({
-    params: PropTypes.shape({
-      id: PropTypes.string.isRequired,
-    }).isRequired
-  }).isRequired,
   onSendComment: PropTypes.func.isRequired,
-  statusAuth: PropTypes.string.isRequired,
-  history: PropTypes.shape({
-    push: PropTypes.func.isRequired,
-  }).isRequired,
+  onInputComment: PropTypes.func.isRequired,
+  onChangeRating: PropTypes.func.isRequired,
+  rating: PropTypes.number.isRequired,
+  comment: PropTypes.string.isRequired,
+  isButtonDisabled: PropTypes.bool.isRequired,
+  readOnly: PropTypes.bool.isRequired,
+  errorText: PropTypes.string.isRequired,
 };
 
-const mapStateToProps = (state) => {
-  const statusAuth = getAuthStatus(state);
+export default AddReview;
 
-  return {
-    statusAuth,
-  };
-};
-
-const mapDispatchToProps = {
-  onSendComment: CommentOperation.sendComment,
-};
-
-export {AddReview};
-export default connect(mapStateToProps, mapDispatchToProps)(AddReview);
